fix: Bypass filters that cannot be applied (#6818)
Use case:

When you enable Workflow for any DocType, it's status is now tracked by
`workflow_field` and not `docstatus`. When you navigate from
Notification dropdown, it applied a `docstatus` filter on the List View
which cannot be applied. This can be ignored and user can apply his
own filters.