fix: Set blog route if title and category are set
Set blog route on client side as soon as title and category are selected.
This behavious is in line with Web Page, where the route is set as soon
as the title is entered. This will ensure blog routes are consistent and
contain category as the part of the route. The user can change the route
ofcourse, but this behaviour is the most common expectation.