test: run patch test to last release and PR
In the patch test, instead of migrating to the current SHA directly.
Migrate to the latest release (with code at the time of release) and
then migrate to last commit in PR.

Why? When directly migrating, the previous patches that would've run
with old code have access to the new code and it's not realistic. This
change brings it a little bit closer to reality. Hence missing failures
that would occur with IRL installations.