fix: Add signature to Communication.content if not already added
This fix adds a signature forcibly if found under the sender's
User.email_signature or default outgoing email account's signature
field.

The previous method of adding a comment into the Email didn't work since
Quill would discard comments before setting them. Adding signatures in
get_formatted_html didn't seem apt since it's used in QueueBuilder to
re-construct the Email before processing the Email Queue. This meant
that the email content that was added in the Communication record would
not be final. Now, we treat the signature as part of the Communication
content.