fix(test): Commit post global_search.sync_global_search
This is an attempt to tackle the flaky test_delete_doc test. Not sure if
this will do anything but the other tests seem to be having a commit
directly after the sync...so following that here.

If this works, the goal would be to find a way to get rid of the commits
altogether. Since the flaky test can't be replicated locally, let's see
if this works !!!!