feat: Show progress of long migration queries (#33660)
Often large databases take several minutes to alter a table, this gives
zero feedback right now and makes it hard to guess where the problem
is if things eventually fail. Was the query slow? Was it making adequate
progress? after how much time it failed?

All of these questions will now be easier to answer. This PR spawns a
new thread everytime we migrate database and...
1. Every 10 seconds check if any command has been running for >10
   seconds from current DB connection.
2. Print details about ongoing command like state, progress, time take
   and query.

The progress detail is often misleading but it's better than nothing.